Implementation of persistence in parkinglot to store payments info related to tickets.
- Improvements on support classes, such as IsNumber
- Implementation of transaction id generation based on the number of payments for a given ticket
- General parsing of app version SupercashMarketplaceNotFoundException
- Adding a new exception named
- Validation of marketplace id (if it exists in the database)
- Updating url for pagarme service
- It not necessary to provide 'items' attribute in the ticket pay request
- Using Long instead of Integer in currency fields

+++ src/main/java/cash/super_/platform/service/pagarme/transactions/models/TransactionResponseSummary.java
@@ -21,7 +21,7 @@ public class TransactionResponseSummary {
      * Paid amount in the transaction
      */
     @JsonProperty(value = "paid_amount")
-    private Integer paidAmount;
+    private Long paidAmount;
 
     /**
      * Supercash internal ID, defined when a request enter for payment.
@@ -52,7 +52,7 @@ public class TransactionResponseSummary {
         return transactionId;
     }
 
-    public Integer getPaidAmount() {
+    public Long getPaidAmount() {
         return paidAmount;
     }
